+++ CHANGELOG.md
@@ -32,6 +32,7 @@ The changelog contains informations about bug fixes, new features or bc breaking
 - `#939` Added ability to toggle block groups and favorize blocks.
 - `#800` Added ability to store the last directory of the filemanager in the user settings via ServiceFoldersDirecotryId service.
 - `#905` Added ability to inject custom ngrest plugin for an appliation specific desinged approach.
+- `#930` Added ability to generate email links in TagParser.
 
 
 1.0.0-beta7 (20.06.2016)